delete-functionality-added
