+++ src/v1/Pages/AddContent/ContentCard.scss
@@ -1,0 +1,38 @@
+@import '../../../Assets/Styles/Variables.scss';
+
+.content-card-container{
+    display: inline-block;
+
+    width: 32%;
+    margin-right: 1%;
+    margin-bottom: 1%;
+
+    border-radius: 10px;
+    position: relative;
+    box-shadow: 0 0 0.5px #888;
+    cursor: pointer;
+}
+
+.content-card-container:hover{
+    box-shadow: 0 0 2px $tabcolor2;
+}
+
+.content-card-edit{
+    position: absolute;
+    right: 5px;
+    top : 4px;
+}
+
+.content-card-img{
+    border-radius: 5px;
+}
+
+
+
+.subcategory-container {
+    border-radius: 10px;
+    max-height: 30rem;
+    overflow-x: hidden;
+    overflow-y: scroll;
+    padding: 10px;
+}
